refac: Minor events clean up
