+++ source/dotnet/wholesale-api/Events/Events.Application/CalculationResultPublishing/ICalculationResultCompletedFactory.cs
@@ -13,10 +13,10 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.CalculationResults.Interfaces.CalculationResults.Model;
+using Energinet.DataHub.Wholesale.Events.Application.CalculationResultPublishing.Model;
 using Energinet.DataHub.Wholesale.Events.Application.IntegrationEventsManagement;
-using Energinet.DataHub.Wholesale.Events.Application.Processes.Model;
 
-namespace Energinet.DataHub.Wholesale.Events.Application.Processes;
+namespace Energinet.DataHub.Wholesale.Events.Application.CalculationResultPublishing;
 
 public interface ICalculationResultCompletedFactory
 {
