+++ source/dotnet/wholesale-api/Events/Events.Application/UseCases/Factories/CompletedBatchFactory.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Batches.Interfaces.Models;
-using Energinet.DataHub.Wholesale.Events.Application.Model;
+using Energinet.DataHub.Wholesale.Events.Application.CompletedBatches;
 using NodaTime.Extensions;
 
 namespace Energinet.DataHub.Wholesale.Events.Application.UseCases.Factories;
