+++ source/dotnet/wholesale-api/Events/Events.Application/UseCases/RegisterCompletedBatchesHandler.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Batches.Interfaces;
-using Energinet.DataHub.Wholesale.Events.Application.Model;
+using Energinet.DataHub.Wholesale.Events.Application.CompletedBatches;
 using Energinet.DataHub.Wholesale.Events.Application.UseCases.Factories;
 
 namespace Energinet.DataHub.Wholesale.Events.Application.UseCases;
