+++ source/dotnet/wholesale-api/Events/Events.Infrastructure/Persistence/EventsDatabaseContext.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using System.Diagnostics.CodeAnalysis;
-using Energinet.DataHub.Wholesale.Events.Application.Model;
+using Energinet.DataHub.Wholesale.Events.Application.CompletedBatches;
 using Energinet.DataHub.Wholesale.Events.Infrastructure.Persistence.Batches;
 using Microsoft.EntityFrameworkCore;
